+++ R/cron.R
@@ -125,65 +125,194 @@ cron2eng <- function(cron = "* * * * *") {
     stop("Argument 'cron' must be a valid cron expression.")
   }
 
-  # Immediate return for special input
-  if (cron == "* * * * *") {
+  # Split each time period to a list element
+  p_list <- as.list(strsplit(cron, " ")[[1]])
 
-    "every minute"
+  # Name elements with text to be used in output for that time period
+  names(p_list) <- c(
+    "minute(s)", "hour(s)", "day(s) of the month",
+    "month(s)", "day(s) of the week"
+  )
 
-  } else {
+  if (stringr::str_detect(p_list[[2]], "\\d")) {
 
-    # Split each time period to a list element
-    p_list <- as.list(strsplit(cron, " ")[[1]])
+    if (stringr::str_detect(p_list[[2]], "/")) {
 
-    # Name elements with text to be used in output for that time period
-    names(p_list) <- c(
-      "minute(s)", "hour(s)", "day(s) of the month",
-      "month(s)", "day(s) of the week"
-    )
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^0(?=/)",  "12AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^1(?=/)",  "1AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^2(?=/)",  "2AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^3(?=/)",  "3AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^4(?=/)",  "4AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^5(?=/)",  "5AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^6(?=/)",  "6AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^7(?=/)",  "7AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^8(?=/)",  "8AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^9(?=/)",  "9AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^10(?=/)", "10AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^11(?=/)", "11AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^12(?=/)", "12PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^13(?=/)", "1PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^14(?=/)", "2PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^15(?=/)", "3PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^16(?=/)", "4PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^17(?=/)", "5PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^18(?=/)", "6PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^19(?=/)", "7PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^20(?=/)", "8PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^21(?=/)", "9PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^22(?=/)", "10PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "^23(?=/)", "11PM")
 
-    # Convert each time period's cron expression to English
-    for (period in names(p_list)) {
+    } else {
 
-      if (p_list[[period]] == "*") {
-
-        # Every valid unit of the time period
-        p_list[[period]] <- paste("every", period)
-
-      } else if (stringr::str_detect(p_list[[period]], "^\\d{1,2}$")) {
-
-        # Single integer value
-        p_list[[period]] <- paste(period, p_list[[period]])
-
-      } else if (stringr::str_detect(p_list[[period]], "^\\d{1,2}-\\d{1,2}$")) {
-
-        # Consecutive units with non-min start and non-max stop value
-        n_split <- strsplit(p_list[[period]], "-")
-        p_list[[period]] <- paste(
-          period, n_split[[1]][1], "to", n_split[[1]][2]
-        )
-
-      } else if (stringr::str_detect(p_list[[period]], "\\d{1,2},")) {
-
-        # Nonconsecutive, irregularly spaced integers
-        n_spaced <- stringr::str_replace_all(p_list[[period]], ",", ", ")
-        p_list[[period]] <- paste(period, n_spaced)
-
-      } else if (stringr::str_detect(p_list[[period]], "\\d{1,2}/\\d{1,2}")) {
-
-        # regularly-spaced integer sequences with specified start and end
-        n_split <- strsplit(p_list[[period]], "/")
-        p_list[[period]] <- paste(
-          "every", n_split[[1]][2], period,
-          "starting from", period, n_split[[1]][1]
-        )
-
-      }
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b0\\b",  "12AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b1\\b",  "1AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b2\\b",  "2AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b3\\b",  "3AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b4\\b",  "4AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b5\\b",  "5AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b6\\b",  "6AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b7\\b",  "7AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b8\\b",  "8AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b9\\b",  "9AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b10\\b", "10AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b11\\b", "11AM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b12\\b", "12PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b13\\b", "1PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b14\\b", "2PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b15\\b", "3PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b16\\b", "4PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b17\\b", "5PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b18\\b", "6PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b19\\b", "7PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b20\\b", "8PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b21\\b", "9PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b22\\b", "10PM")
+      p_list[[2]] <- stringr::str_replace(p_list[[2]], "\\b23\\b", "11PM")
 
     }
 
-    # Collate sentences for each time period
-    paste(p_list, collapse = "; ")
+  }
+
+  if (stringr::str_detect(p_list[[4]], "\\d")) {
+
+    if (stringr::str_detect(p_list[[4]], "/")) {
+
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "^1(?=/)", "January")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "^2(?=/)", "February")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "^3(?=/)", "March")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "^4(?=/)", "April")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "^5(?=/)", "May")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "^6(?=/)", "June")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "^7(?=/)", "July")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "^8(?=/)", "August")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "^9(?=/)", "September")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "^10(?=/)", "October")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "^11(?=/)", "November")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "^12(?=/)", "December")
+
+    } else {
+
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "\\b1\\b", "January")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "\\b2\\b", "February")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "\\b3\\b", "March")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "\\b4\\b", "April")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "\\b5\\b", "May")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "\\b6\\b", "June")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "\\b7\\b", "July")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "\\b8\\b", "August")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "\\b9\\b", "September")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "\\b10\\b", "October")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "\\b11\\b", "November")
+      p_list[[4]] <- stringr::str_replace(p_list[[4]], "\\b12\\b", "December")
+
+    }
 
   }
+
+  if (stringr::str_detect(p_list[[5]], "\\d")) {
+
+    if (stringr::str_detect(p_list[[5]], "/")) {
+
+      p_list[[5]] <- stringr::str_replace(p_list[[5]], "0(?=/)", "Sunday")
+      p_list[[5]] <- stringr::str_replace(p_list[[5]], "1(?=/)", "Monday")
+      p_list[[5]] <- stringr::str_replace(p_list[[5]], "2(?=/)", "Tuesday")
+      p_list[[5]] <- stringr::str_replace(p_list[[5]], "3(?=/)", "Wednesday")
+      p_list[[5]] <- stringr::str_replace(p_list[[5]], "4(?=/)", "Thursday")
+      p_list[[5]] <- stringr::str_replace(p_list[[5]], "5(?=/)", "Friday")
+      p_list[[5]] <- stringr::str_replace(p_list[[5]], "6(?=/)", "Saturday")
+
+    } else {
+
+      p_list[[5]] <- stringr::str_replace(p_list[[5]], "0", "Sunday")
+      p_list[[5]] <- stringr::str_replace(p_list[[5]], "1", "Monday")
+      p_list[[5]] <- stringr::str_replace(p_list[[5]], "2", "Tuesday")
+      p_list[[5]] <- stringr::str_replace(p_list[[5]], "3", "Wednesday")
+      p_list[[5]] <- stringr::str_replace(p_list[[5]], "4", "Thursday")
+      p_list[[5]] <- stringr::str_replace(p_list[[5]], "5", "Friday")
+      p_list[[5]] <- stringr::str_replace(p_list[[5]], "6", "Saturday")
+
+    }
+
+  }
+
+  # Convert each time period's cron expression to English
+  for (p in names(p_list)) {
+
+    if (p_list[[p]] == "*") {
+
+      # Every valid unit of the time period
+      p_list[[p]] <- paste("every", p)
+
+    } else if (stringr::str_detect(p_list[[p]], "^\\d{1,2}$|^\\w{1,}$")) {
+
+      # Single integer value
+      p_list[[p]] <- paste(p, p_list[[p]])
+
+    } else if (stringr::str_detect(p_list[[p]], "^\\w{1,}-\\w{1,}$")) {
+
+      # Consecutive units with non-min start and non-max stop value
+      n_split <- strsplit(p_list[[p]], "-")
+      p_list[[p]] <- paste(
+        p, n_split[[1]][1], "to", n_split[[1]][2]
+      )
+
+    } else if (stringr::str_detect(p_list[[p]], "^\\w{1,},")) {
+
+      # Nonconsecutive, irregularly spaced integers
+      n_units <- stringr::str_split(p_list[[p]], ",")[[1]]
+      p_list[[p]] <- paste(p, knitr::combine_words(n_units))
+
+    } else if (stringr::str_detect(p_list[[p]], "^\\w{1,}/\\w{1,}$")) {
+
+      # regularly-spaced integer sequences with specified start and end
+      n_split <- strsplit(p_list[[p]], "/")
+      p_list[[p]] <- paste0(
+        "every ", n_split[[1]][2], " ", p,
+        " starting from ",
+        ifelse(p %in% c("minute(s)", "day(s)"), paste0(p, " "), ""),
+        n_split[[1]][1]
+      )
+
+    }
+
+  }
+
+  # Collate sentences for each time period
+  # paste0(paste(p_list, collapse = "; "), ".")
+  # paste0(
+  #   p_list[[1]], " of ", p_list[[2]],
+  #   "; on ", p_list[[3]], ", of ", p_list[[4]],
+  #   "; incuding ", p_list[[5]], "."
+  # )
+  cat(
+    paste0("Cron string '", cron, "' means:"),
+    paste0("  • ", p_list[[1]]),
+    paste0("  • ", p_list[[2]]),
+    paste0("  • ", p_list[[3]]),
+    paste0("  • ", p_list[[4]]),
+    paste0("  • ", p_list[[5]]),
+    sep = "\n"
+  )
 
 }